vulnerable_app/requirements.txt
Flask==2.3.2
psycopg2_binary==2.9.6
